chore: move fake client into test file
