+++ tests/unit/conftest.py
@@ -12,14 +12,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from typing import Any
-
-from aiohttp import web
 from mocks import (
-    connectionInfo,
     FakeCredentials,
     FakeInstance,
-    generateClientCertificate,
 )
 import pytest
 
@@ -27,16 +22,6 @@ import pytest
 @pytest.fixture
 def credentials() -> FakeCredentials:
     return FakeCredentials()
-
-
-@pytest.fixture
-async def client(aiohttp_client: Any) -> Any:
-    app = web.Application()
-    metadata_uri = "/v1beta/projects/test-project/locations/test-region/clusters/test-cluster/instances/test-instance/connectionInfo"
-    app.router.add_get(metadata_uri, connectionInfo)
-    client_cert_uri = "/v1beta/projects/test-project/locations/test-region/clusters/test-cluster:generateClientCertificate"
-    app.router.add_post(client_cert_uri, generateClientCertificate)
-    return await aiohttp_client(app)
 
 
 @pytest.fixture
