+++ tests/unit/mocks.py
@@ -13,10 +13,8 @@
 # limitations under the License.
 
 from datetime import datetime, timedelta
-import json
-from typing import Any, Callable, Tuple
+from typing import Callable, Tuple
 
-from aiohttp import web
 from cryptography import x509
 from cryptography.hazmat.primitives import hashes, serialization
 from cryptography.hazmat.primitives.asymmetric import rsa
@@ -52,25 +50,6 @@ class FakeCredentials:
         is not expired.
         """
         return self.token is not None and not self.expired
-
-
-async def connectionInfo(request: Any) -> web.Response:
-    response = {
-        "ipAddress": "127.0.0.1",
-        "instanceUid": "123456789",
-    }
-    return web.Response(content_type="application/json", body=json.dumps(response))
-
-
-async def generateClientCertificate(request: Any) -> web.Response:
-    response = {
-        "pemCertificate": "This is the client cert",
-        "pemCertificateChain": [
-            "This is the intermediate cert",
-            "This is the root cert",
-        ],
-    }
-    return web.Response(content_type="application/json", body=json.dumps(response))
 
 
 def generate_cert(
